Improve collection search
- Decouple collection queries from presentation.
- Decouple query criteria from each other

[#167062277]

+++ search-service/src/main/kotlin/com/boclips/search/service/infrastructure/contract/AbstractInMemoryFake.kt
@@ -18,8 +18,8 @@ abstract class AbstractInMemoryFake<QUERY : SearchQuery<METADATA>, METADATA> :
         val idsMatching = idsMatching(index, searchRequest.query)
 
         return sort(idsMatching, searchRequest.query)
-            .drop(searchRequest.startIndex)
-            .take(searchRequest.windowSize)
+            .drop(searchRequest.startIndex.toInt())
+            .take(searchRequest.windowSize.toInt())
     }
 
     private fun sort(ids: List<String>, query: QUERY): List<String> {
